Refactor api module
Split out all the endpoint calls to files per category and import them
back into the main api module.
