+++ frontend2/src/api/authentication.ts
@@ -1,0 +1,20 @@
+import axios from 'axios';
+import { apiUrl } from '@/env';
+
+export const authentication = {
+  async logInGetToken(username: string, password: string) {
+    const params = new URLSearchParams();
+    params.append('username', username);
+    params.append('password', password);
+    return axios.post(`${apiUrl}/api/v1/login/access-token`, params);
+  },
+  async passwordRecovery(email: string) {
+    return axios.post(`${apiUrl}/api/v1/password-recovery/${email}`);
+  },
+  async resetPassword(password: string, token: string) {
+    return axios.post(`${apiUrl}/api/v1/reset-password/`, {
+      NewPassword: password,
+      token,
+    });
+  },
+};
