+++ frontend2/src/api/headers.ts
@@ -1,0 +1,7 @@
+export function authHeaders(token: string) {
+  return {
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
+  };
+}
